+++ src/core/entities/todo/todo.ts
@@ -2,7 +2,9 @@ import { TodoStatus } from './enum';
 
 export class Todo {
   public name: string;
+
   public status: TodoStatus;
+
   public user: string;
 
   constructor(props: Partial<Todo>) {