+++ src/interface/cli/create-todo/create-todo.interface.ts
@@ -1,0 +1,5 @@
+import { CliResult, CliLine } from '..';
+
+export interface ICreateTodoCommand {
+  cmd(line: CliLine): CliResult | void;
+}