+++ src/interface/cli/index.ts
@@ -1,3 +1,3 @@
-export * from './cli';
+export * from './create-todo/create-todo.interface';
 export * from './cli.types';
 export * from './create-todo/create-todo';