+++ src/main/modules/cli/cli.spec.ts
@@ -1,0 +1,24 @@
+import { assert } from 'chai';
+import sinon from 'sinon';
+import { CliModule } from './cli';
+
+describe('Cli', () => {
+  describe('start', () => {
+    it('Should start program command with succeffully', () => {
+      const program = {
+        on: sinon.fake.returns(undefined),
+      };
+      const cli = new CliModule(
+        {
+          // @ts-ignore
+          program,
+        },
+        program
+      );
+
+      cli.start();
+
+      assert(program.on.calledOnce);
+    });
+  });
+});