+++ src/main/modules/index.ts
@@ -1,0 +1,2 @@
+export * from './modules.interface';
+export * from './cli/cli';