+++ src/presentation/cli/cli.ts
@@ -1,6 +1,0 @@
-import { CliResult } from '.';
-import { CliLine } from '.';
-
-export abstract class Cli {
-  abstract cmd(line: CliLine): CliResult | void;
-}